Added CloudSettings
